requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.11
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.3.0
base58==2.1.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==44.0.0
did-resolver==0.0.3
Django==5.2.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
ecdsa==0.19.1
frozenlist==1.6.2
geographiclib==2.0
geojson==3.2.0
geopy==2.4.1
grpcio==1.68.1
grpcio-tools==1.68.1
hiero-sdk-python==0.1.1
hiero_did_sdk_python==0.1.0
idna==3.10
multidict==6.4.4
numpy==2.3.0
pillow==11.2.1
propcache==0.3.1
protobuf==5.28.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
PyJWT==2.9.0
pyproj==3.7.1
python-dotenv==1.0.1
python-jose==3.5.0
requests==2.32.3
rest-framework-simplejwt==0.0.2
rsa==4.9.1
setuptools==80.9.0
shapely==2.1.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.0
urllib3==2.4.0
yarl==1.20.0
zstandard==0.23.0
